+++ modules/main/src/main/java/com/annimon/ownlang/modules/okhttp/HttpClientValue.java
@@ -8,6 +8,7 @@ import okhttp3.Response;
 import okhttp3.WebSocket;
 import okhttp3.WebSocketListener;
 import okio.ByteString;
+import java.util.Objects;
 
 public class HttpClientValue extends MapValue {
 
@@ -117,6 +118,7 @@ public class HttpClientValue extends MapValue {
                 }
             }
         });
+        Objects.requireNonNull(ws);
         return new CallValue(client.newCall(request));
     }
 }